Add IDs to ancestor-descendant tables
To possibly allow re-inserts (deletes followed by inserts in one query)
of rows that keep ancestor_id and descendant_id unchanged. Previously
the fact that these columns constituted a primary key proved to be an
obstacle when performing delete and insert in one query via CTEs.

This change is being made with '_move_file()' in mind.